requirements.txt
pandas==1.2.3
plotly==4.14.3
plotly-express==0.4.1
scikit-surprise==1.1.1